apps/worker/requirements.txt
rq==1.16.0
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
python-dotenv==1.0.0
